Last update (released in 01/09/2021 at server of current company[KWS])

+++ TeamCalendar/src/main/java/keyware/unyu/seisan/DAO/UserDAO.java
@@ -79,8 +79,8 @@ public class UserDAO {
 	}
 	
 	//TeamScheduleに登録した番号読み込む
-	public int getTeamSchedule() {
-		int result = 0;
+	public long getTeamSchedule() {
+		long result = 0;
 		UserMapper mapper = session.getMapper(UserMapper.class);
 		try {
 			result = mapper.getTeamSchedule();
